Increase size of some `struct fcb` members
- Moving the sector_cnt and scratch_cnt members to a uint16_t allows for a counts > 255.
- Moving the version to a uint16_t allows for a more flexible scheme of versioning.

Component: Flash Circular Buffer

+++ fs/fcb/include/fcb/fcb.h
@@ -52,7 +52,7 @@ struct fcb_entry {
 struct fcb {
     /* Caller of fcb_init fills this in */
     uint32_t f_magic;		/* As placed on the disk */
-    uint8_t f_version;  	/* Current version number of the data */
+    uint16_t f_version;  	/* Current version number of the data */
     uint16_t f_sector_cnt;	/* Number of elements in sector array */
     uint16_t f_scratch_cnt;	/* How many sectors should be kept empty */
     struct flash_area *f_sectors; /* Array of sectors, must be contiguous */
